requirements.txt

# Database
psycopg[binary]==3.1.18

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# AI/ML
openai==1.3.0
paddlepaddle==2.6.2
paddleocr==2.7.0

# Document Processing
PyMuPDF==1.20.0
pdfplumber==0.10.3
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0